Moved models into models folder, removed old versions of tasks
