+++ app/pages/tasks/task-detail/task-detail.component.html
@@ -3,7 +3,7 @@
     </ActionItem>
 </ActionBar>
 <StackLayout class="container">
-    <!-- <Label class="duration-heading" text="{{ task.duration }}" textAlignment="center"></Label> -->
+    <Label class="duration-heading" text="{{ task.duration }}" textAlignment="center"></Label>
     <StackLayout class="hr-light m-10"></StackLayout>
     <Label class="description-body" text="{{ task.name }}" textWrap="true"></Label>
     <Label class="note-heading" text="Note:" textAlignment="center"></Label>
